# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: isachore API
  description: |
    Chore management for households.

    ## Authenticating

    Two transports, either of which carries a session:

    - the httpOnly `isachore_token` cookie, which a browser sends on its own, and
    - the same opaque token as an `Authorization: Bearer` header, for clients with no cookie jar.

    Two ways to open one, and nothing downstream can tell them apart. `POST /api/v1/auth/login`
    takes an email and password, and answers `two_factor_required` instead of a session when the
    account has TOTP enabled - finish that with `POST /api/v1/auth/verify-2fa`. Where single
    sign-on is configured, `GET /api/v1/auth/oidc/start` is the other way in, and on a server
    with `OIDC_ONLY` set it is the only one: password login answers 403 there.
    `GET /api/v1/auth/methods` says which of the two this server offers, without a session.

    There is no self-registration: administrators create accounts, and the first administrator
    comes from the `init` CLI command.

    ## The CSRF header

    A request authenticated **by cookie** that uses an unsafe method (POST, PATCH, PUT, DELETE)
    must also carry a non-empty `X-CSRF-Token` header, or it is refused with 403 before it
    reaches the route. Any value will do: this is a custom-header defence over `SameSite=Lax`,
    so what matters is that a cross-site form cannot set the header at all.

...
  version: 0.1.0
paths:
...
  /api/v1/tags:
    get:
      tags:
        - tags
      summary: List Tags
      operationId: list_tags_api_v1_tags_get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: sort_by
          in: query
          required: false
          schema:
            enum:
              - id
              - name
              - created_at
            type: string
            default: name
            title: Sort By
        - name: sort_dir
          in: query
          required: false
          schema:
            enum:
              - asc
              - desc
            type: string
            default: asc
            title: Sort Dir
        - name: household_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 1
              - type: 'null'
            title: Household Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TagRead_'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: 'Tags are an organiser surface, and both halves of this are about that role rather than membership: no household with the id given that you ORGANISE - a deputy or helper of it reaches this too - or, when no id is given, you organise none at all, so there is nothing to fall back to.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - tags
      summary: Create Tag
      operationId: create_tag_api_v1_tags_post
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No household with this id that you are a member of.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: That household already has a tag with this name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tags/{tag_id}:
    get:
      tags:
        - tags
      summary: Get Tag
      operationId: get_tag_api_v1_tags__tag_id__get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: integer
            title: Tag Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No tag with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - tags
      summary: Update Tag
      operationId: update_tag_api_v1_tags__tag_id__patch
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: integer
            title: Tag Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No tag with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: That household already has a tag with this name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - tags
      summary: Delete Tag
      operationId: delete_tag_api_v1_tags__tag_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: tag_id
          in: path
          required: true
          schema:
            type: integer
            title: Tag Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No tag with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/chores/{chore_id}:
    get:
      tags:
        - chores
      summary: Get Chore
      operationId: get_chore_api_v1_chores__chore_id__get
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: chore_id
          in: path
          required: true
          schema:
            type: integer
            title: Chore Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No chore with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - chores
      summary: Update Chore
      operationId: update_chore_api_v1_chores__chore_id__patch
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: chore_id
          in: path
          required: true
          schema:
            type: integer
            title: Chore Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreRead'
        '400':
          description: 'An assignee, a tag or the named current assignee does not belong where it must: assignees and tags have to be from the chore''s own household, and the current assignee has to be one of the chore''s assignees.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No chore with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Somebody completed or skipped this chore while the edit was in flight, so the rescheduled occurrence collided with the closed one. Retrying is safe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - chores
      summary: Delete Chore
      operationId: delete_chore_api_v1_chores__chore_id__delete
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: chore_id
          in: path
          required: true
          schema:
            type: integer
            title: Chore Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: A member of the household, but the role held there does not reach what this route requires.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          description: No chore with this id in any household you belong to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssignmentType:
      type: string
      enum:
        - manual
        - alphabetical
        - random
        - least_done
      title: AssignmentType
...
    ChoreHouseholdRead:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        timezone:
          type: string
          title: Timezone
      type: object
      required:
        - id
        - name
        - timezone
      title: ChoreHouseholdRead
      description: |-
        The household a chore belongs to, embedded in both chore reads (it sits on
        ChoreReadBase) so the list column and the read-only edit header can show its name
        without a second request.

        Also embedded in `DueChoreRead`, `UnscheduledChoreRead`, `HistoryEntryRead` and the
        History filter options, which is what makes `timezone` here reach every household-scoped
        surface at once.
...
    ChoreRead:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        start_date:
          anyOf:
            - type: string
              format: date
            - type: 'null'
          title: Start Date
        repeats:
          $ref: '#/components/schemas/RepeatPeriod'
        assignment_type:
          $ref: '#/components/schemas/AssignmentType'
        turn_length:
          type: integer
          title: Turn Length
        repeat_interval:
          type: integer
          title: Repeat Interval
        weekdays:
          anyOf:
            - items:
                type: integer
              type: array
            - type: 'null'
          title: Weekdays
        created_at:
          type: string
          format: date-time
          title: Created At
        household:
          $ref: '#/components/schemas/ChoreHouseholdRead'
        assignees:
          items:
            $ref: '#/components/schemas/HouseholdMemberRead'
          type: array
          title: Assignees
        current_assignee:
          anyOf:
            - $ref: '#/components/schemas/HouseholdMemberRead'
            - type: 'null'
        tags:
          items:
            $ref: '#/components/schemas/TagRead'
          type: array
          title: Tags
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
      type: object
      required:
        - id
        - title
        - start_date
        - repeats
        - assignment_type
        - turn_length
        - repeat_interval
        - weekdays
        - created_at
        - household
        - assignees
        - tags
        - description
      title: ChoreRead
      description: |-
        A single chore: create, `GET /chores/{id}` and patch. Open to every role, since the
        description dialog on My Chores and Unscheduled fetches it.
    ChoreUpdate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          anyOf:
            - type: string
              maxLength: 20000
            - type: 'null'
          title: Description
        start_date:
          anyOf:
            - type: string
              format: date
            - type: 'null'
          title: Start Date
        repeats:
          $ref: '#/components/schemas/RepeatPeriod'
        assignment_type:
          $ref: '#/components/schemas/AssignmentType'
        turn_length:
          type: integer
          minimum: 1
          title: Turn Length
          default: 1
        repeat_interval:
          type: integer
          maximum: 365
          minimum: 1
          title: Repeat Interval
          default: 1
        weekdays:
          anyOf:
            - items:
                type: integer
                maximum: 6
                minimum: 0
              type: array
            - type: 'null'
          title: Weekdays
        assignee_ids:
          items:
            type: integer
          type: array
          title: Assignee Ids
        current_assignee_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Current Assignee Id
        clear_current_assignee:
          type: boolean
          title: Clear Current Assignee
          default: false
        tag_ids:
          items:
            type: integer
          type: array
          title: Tag Ids
      type: object
      required:
        - title
        - repeats
        - assignment_type
      title: ChoreUpdate
      description: |-
        Full replace of an editable chore. The household is fixed at creation and
        intentionally not editable here.
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
        - detail
      title: ErrorDetail
      description: |-
        The body FastAPI sends for a raised `HTTPException`.

        Declared so a refusal can be documented as carrying its explanation rather than as
        bodyless, which is what a route with only a `description` claims. Every hand-raised
        HTTPException in this app answers in this shape; only the routes that declare it say
        so, and wiring that up everywhere is the job of the security-scheme step in README's
        todo rather than of any single endpoint.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HouseholdMemberRead:
      properties:
        id:
          type: integer
          title: Id
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
      type: object
      required:
        - id
        - first_name
        - last_name
      title: HouseholdMemberRead
...
    Page_TagRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TagRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
        - items
        - total
        - page
        - page_size
      title: Page[TagRead]
...
    RepeatPeriod:
      type: string
      enum:
        - manual
        - daily
        - weekly
        - monthly
        - yearly
      title: RepeatPeriod
...
    TagCreate:
      properties:
        household_id:
          type: integer
          title: Household Id
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        color:
          type: string
          pattern: ^#[0-9a-fA-F]{6}$
          title: Color
      type: object
      required:
        - household_id
        - name
        - color
      title: TagCreate
    TagRead:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        color:
          type: string
          title: Color
      type: object
      required:
        - id
        - name
        - color
      title: TagRead
    TagUpdate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        color:
          type: string
          pattern: ^#[0-9a-fA-F]{6}$
          title: Color
      type: object
      required:
        - name
        - color
      title: TagUpdate
      description: |-
        Full replace of an editable tag. The household is fixed at creation and
        intentionally not editable here (mirrors ChoreUpdate).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: The httpOnly session cookie, set by `POST /api/v1/auth/login` or by the single sign-on callback. A browser sends it automatically; an unsafe method additionally needs the `X-CSRF-Token` header.
      in: cookie
      name: isachore_token
    bearerToken:
      type: http
      description: 'The same opaque session token as an `Authorization: Bearer` header, for clients with no cookie jar. Exempt from the `X-CSRF-Token` requirement.'
      scheme: bearer
    parkedAdminCookie:
      type: apiKey
      description: An impersonating administrator's own session, parked while they act as somebody else. `POST /api/v1/auth/stop-impersonating` is the only route it *authenticates* - several others read it to attribute an action to the real operator, and logout reads it to end both sessions.
      in: cookie
      name: isachore_admin_token